requirements.txt
absl-py==1.0.0
aniso8601==9.0.1
attrs==21.4.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cvxopt==1.2.7
cycler==0.11.0
Flask==2.0.2
Flask-Autodoc==0.1.2
flask-restplus==0.13.0
fonttools==4.28.1
google-auth==2.3.3
google-auth-oauthlib==0.4.6
grpcio==1.42.0
idna==3.3
importlib-metadata==4.8.2
itsdangerous==2.0.1
Jinja2==3.0.3
jsonschema==4.4.0
kaggle==1.5.12
kiwisolver==1.3.2
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.0
minio==7.1.1
numpy==1.21.4
oauthlib==3.1.1
opencv-python==4.5.5.62
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.6
pyrsistent==0.18.1
python-dateutil==2.8.2
python-mnist==0.7
python-slugify==5.0.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
setuptools-scm==6.3.2
six==1.16.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
text-unidecode==1.3
tomli==1.2.2
# torch==1.10.1+cu113
# torchaudio==0.10.1+cu113
# torchvision==0.11.2+cu113
tqdm==4.62.3
typing_extensions==4.0.0
urllib3==1.26.7
utils==1.0.1
Werkzeug==2.0.2
zipp==3.6.0
